requirements.txt
pandas
scikit-learn
torch==2.2.2  
torchvision
matplotlib
Pillow  
scipy  
torchsummary